+++ app/src/main/java/com/example/woowagithubrepositoryapp/network/GithubClient.kt
@@ -7,32 +7,29 @@ import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
 class GithubClient {
-    fun<T> generate(api: Class<T>):T = Retrofit.Builder()
+    fun <T> generate(api: Class<T>): T = Retrofit.Builder()
         .baseUrl(App.HOST)
         .addConverterFactory(GsonConverterFactory.create())
         .client(httpClient())
         .build()
         .create(api)
 
-    fun<T> generateRefreshClient(api: Class<T>):T = Retrofit.Builder()
+    fun <T> generateRefreshClient(api: Class<T>): T = Retrofit.Builder()
         .baseUrl(App.AUTH_HOST)
         .addConverterFactory(GsonConverterFactory.create())
         .client(refreshClient())
         .build()
         .create(api)
 
-
     private fun httpClient() =
         OkHttpClient.Builder().apply {
             addInterceptor(httpLoggingInterceptor())
             addInterceptor(TokenInterceptor())
-            authenticator(TokenAuthenticator())
         }.build()
 
     private fun refreshClient() =
         OkHttpClient.Builder().apply {
             addInterceptor(httpLoggingInterceptor())
-            addInterceptor(RefreshTokenInterceptor())
         }.build()
 
     private fun httpLoggingInterceptor() =
