+++ app/src/main/java/com/example/woowagithubrepositoryapp/network/TokenInterceptor.kt
@@ -1,6 +1,8 @@
 package com.example.woowagithubrepositoryapp.network
 
+import com.example.woowagithubrepositoryapp.ui.auth.LoginActivity
 import com.example.woowagithubrepositoryapp.utils.Prefs
+import com.example.woowagithubrepositoryapp.utils.clearTasksAndStartActivity
 import okhttp3.Interceptor
 import okhttp3.Response
 
@@ -13,7 +15,13 @@ class TokenInterceptor : Interceptor {
             method(original.method, original.body)
         }.build()
 
-        return chain.proceed(request)
-    }
+        val response = chain.proceed(request)
 
+        if (response.code == 401) {
+            Prefs.accessToken = ""
+            clearTasksAndStartActivity<LoginActivity>()
+        }
+
+        return response
+    }
 }
